Fix the scraping of the air monitor stations
NYC DEC recently switched the entire url and webpage for where they
store and display air monitor data. Unfortunately, the new page has a
bunch of ASP and javascript, so I have implemented a selenium approach
to scraping the data. I need to configure the server properly to handle
this (display and Firefox).